requirements.txt
gradio=4.44.1
open-interpreter